requirements.txt
streamlit
pandas
numpy
plotly
spacy
wordcloud
fpdf
Pillow
keybert
sentence-transformers
torch
transformers
